cmenu: use getscreensize to access number of rows/columns
The simple menu system already uses getscreensize. Another step towards
unification.

Impact: exported getnumcols and getnumrows are now deprecated.

+++ com32/cmenu/libmenu/menu.h
@@ -24,6 +24,7 @@
 #include "syslnx.h"
 #include "scancodes.h"
 #include <string.h>
+#include <unistd.h>
 
 // TIMEOUT PARAMETERS
 /* If no key is pressed within TIMEOUTNUMSTEPS * TIMEOUTSTEPSIZE milliseconds
@@ -214,8 +215,8 @@ typedef struct s_menusystem {
     uchar shadowattr;
     uchar statline;
     uchar menupage;
-    uchar maxrow, minrow, numrows;	// Number of rows in the window
-    uchar maxcol, mincol, numcols;	// Number of columns in the window
+    int maxrow, minrow, numrows;	// Number of rows in the window
+    int maxcol, mincol, numcols;	// Number of columns in the window
 
     // Menu box look
     boxtype menubt;		// What type of boxes should be drawn
